Update mutate.js
It was simple but tricky

Three variables are declared. The fElement, sElement variables stores the return of converting the provided array of index 0 and 1 respectively to lower case. While result stores Boolean true value if the index return a positive value greater than or equals 0.

A for loop is then run to iterate through the sElement array element and check if itexist in fElement.